add test expand test on cluster resource tests
